[IMP] website_sale(_stock): move logic
Functions moved from website_sale_stock to website_sale:
- _is_sold_out
- _website_show_quick_add
- _to_markup_data
- part of _get_additionnal_combination_info
- _get_cart_qty
- _get_common_product_lines

Views and scripts moved from website_sale_stock to website_sale:
- wishlist_templates.xml
- part of website_sale_stock.product_availability
- availability_messages div
- parts of product_page.js

task 5418128
