+++ addons/website_sale/models/website.py
@@ -1203,3 +1203,17 @@ class Website(models.Model):
         :rtype: dict
         """
         return json_scriptsafe.dumps(self._prepare_ecommerce_store_markup_data(), indent=2)
+
+    def _get_product_available_qty(self, product, **_kwargs):
+        """Give the available quantity of a given product.
+
+        NB: this method is only meant to be used on the shop before the checkout.
+        For checkout steps, please use `cart._get_free_qty` instead to consider
+        the chosen warehouse for delivery (website_sale_collect).
+
+        :param product: product.product record
+        :param dict kwargs: unused parameters, available for overrides
+        :return: available quantity
+        :rtype: float
+        """
+        return product.qty_available - product.outgoing_qty
