+++ addons/website_sale/templates/checkout/cart_templates.xml
@@ -227,6 +227,7 @@
                 />
             </t>
         </div>
+        <div class="availability_messages"/>
     </template>
 
     <template id="website_sale.cart_lines_price" name="Shopping Cart Line Price">
