+++ addons/website_sale/templates/product_page_templates.xml
@@ -549,6 +549,7 @@
     </template>
 
     <template id="website_sale.cta_wrapper" name="CTA Wrapper">
+        <div class="availability_messages o_not_editable"/>
         <div
             id="o_wsale_cta_wrapper"
             t-att-class="'d-flex flex-wrap align-items-center gap-2 mb-4' +
