+++ addons/website_sale_stock/__manifest__.py
@@ -18,7 +18,6 @@ Then it can be made specific at the product level.
         "views/sale_order_views.xml",
         "views/stock_picking_views.xml",
         "views/website_sale_stock_templates.xml",
-        "views/wishlist_templates.xml",
         "data/template_email.xml",
         "data/ir_cron_data.xml",
         # Wizards
