+++ addons/website_sale_stock/models/website.py
@@ -9,7 +9,8 @@ class Website(models.Model):
     warehouse_id = fields.Many2one("stock.warehouse", string="Warehouse")
 
     def _get_product_available_qty(self, product, **_kwargs):
-        """Give the available quantity of a given product.
+        """ Override of _get_product_available_qty in website_sale module
+        Give the available quantity of a given product.
 
         NB: this method is only meant to be used on the shop before the checkout.
         For checkout steps, please use `cart._get_free_qty` instead to consider
