+++ addons/website_sale_stock/views/website_sale_stock_templates.xml
@@ -9,9 +9,6 @@
         <xpath expr="//input[@type='text'][hasclass('quantity')]" position="attributes">
             <attribute name="t-att-data-max">line._get_max_line_qty()</attribute>
         </xpath>
-        <xpath expr="//div[contains(@t-attf-class, 'css_quantity')]" position="after">
-            <div class="availability_messages"/>
-        </xpath>
     </template>
 
     <template
@@ -20,7 +17,6 @@
         priority="4"
     >
         <xpath expr="//div[@id='o_wsale_cta_wrapper']" position="before">
-            <div class="availability_messages o_not_editable"/>
             <input id="wsale_user_email" type="hidden" t-att-value="user_email"/>
         </xpath>
     </template>
