--- addons/website_sale_stock/views/wishlist_templates.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <template id="product_wishlist" inherit_id="website_sale.product_wishlist">
-        <xpath expr="//t[@t-call='website_sale.products_item']" position="before">
-            <t t-set="product_extra_information">
-                <span
-                    t-if="wish.product_id._is_sold_out() and not wish.product_id.allow_out_of_stock_order"
-                    class="o_wsale_stock_badge badge text-bg-danger align-self-start align-top"
-                >
-                    <i class="fa fa-circle text-danger me-1" aria-hidden="true"/>
-                    Out of stock
-                </span>
-            </t>
-        </xpath>
-        <xpath expr="//t[@t-call='website_sale.products_item']" position="attributes">
-            <attribute name="product_extra_information">product_extra_information</attribute>
-        </xpath>
-        <button name="add_to_cart" position="attributes">
-            <attribute name="t-att-disabled">wish.product_id._is_sold_out()</attribute>
-        </button>
-    </template>
-</odoo>
